글 삭제 할 때 로그인한 사람이 작성자인지 판단

+++ pages/api/delete.js
@@ -1,9 +1,20 @@
 import {connectDB} from "@/util/database";
 import {ObjectId} from "mongodb";
+import {getServerSession} from "next-auth";
+import {authOptions} from "@/pages/api/auth/[...nextauth]";
 
 export default async function handler(request, response) {
     if (request.method === "POST") {
         try {
+            let session = await getServerSession(request, response, authOptions);
+            if (session === null) {
+                response.status(401).json("로그인 하셈 ㅅㄱ");
+                return;
+            }
+            if(session.user.email !== request.body.regId) {
+                response.status(403).json("너가 쓴거 아님 ㅅㄱ");
+                return;
+            }
             const db = (await connectDB).db("forum");
             let result = await db.collection('post').deleteOne({_id: new ObjectId(request.body.id)});
             if (result.deletedCount > 0) {
